+++ test/system/validations_test.rb
@@ -17,7 +17,7 @@ class ValidationsTest < ApplicationSystemTestCase
   end
 
   test "validates fields on the server" do
-    visit new_message_path
+    visit new_message_path(disableSubmitWhenInvalid: true)
 
     within_fieldset "Validate" do
       select "published", from: "Status"
@@ -25,13 +25,14 @@ class ValidationsTest < ApplicationSystemTestCase
       fill_in "Content", with: "not empty"
       click_on "Create Message"
 
-      assert_field "Subject", valid: false, validation_message: "is reserved"
-      assert_button "Create Message", disabled: false
+      assert_field "Subject", valid: false, validation_message: "is reserved", focused: true
+      assert_button "Create Message", disabled: true
+      assert_button "Skip Validations", disabled: false
     end
   end
 
-  test "focuses first invalid field on submission" do
-    visit new_message_path
+  test "focuses first invalid field when connected with multiple invalid fields" do
+    visit new_message_path(disableSubmitWhenInvalid: false)
 
     within_fieldset "Validate" do
       click_on "Create Message"
@@ -39,6 +40,19 @@ class ValidationsTest < ApplicationSystemTestCase
       assert_field "Status", valid: false, focused: true
       assert_field "Subject", valid: false, focused: false
       assert_field "Content", valid: false, focused: false
+    end
+  end
+
+  test "disables submit button when server-rendered response is invalid" do
+    visit new_message_path(disableSubmitWhenInvalid: true)
+
+    within_fieldset "Validate" do
+      click_on "Skip Validations"
+
+      assert_field "Status", valid: false, focused: true
+      assert_field "Subject", valid: false, focused: false
+      assert_field "Content", valid: false, focused: false
+      assert_button "Create Message", disabled: true
     end
   end
 
@@ -60,12 +74,14 @@ class ValidationsTest < ApplicationSystemTestCase
     visit new_message_path(disableSubmitWhenInvalid: true)
 
     within_fieldset "Validate" do
-      assert_button "Create Message", disabled: false
+      assert_button "Create Message", disabled: true
+      assert_button "Skip Validations", disabled: false
 
       tab_until_focused :field, "Content"
       send_keys "valid"
 
       assert_button "Create Message", disabled: true
+      assert_button "Skip Validations", disabled: false
       assert_field "Status", validation_message: "can't be blank"
       assert_field "Subject", validation_message: "can't be blank"
 
@@ -73,10 +89,34 @@ class ValidationsTest < ApplicationSystemTestCase
       fill_in("Subject", with: "valid").then { tab_until_focused :field, "Content" }
 
       assert_button "Create Message", disabled: false
+      assert_button "Skip Validations", disabled: false
       assert_no_text "can't be blank"
       assert_no_field validation_message: "Please select an item in the list."
       assert_no_field validation_message: "can't be blank"
     end
+  end
+
+  test "clears validation state when submission is valid" do
+    visit new_message_path(disableSubmitWhenInvalid: false)
+
+    within_fieldset "Validate" do
+      click_on "Create Message"
+      assert_field valid: false, count: 3
+    end
+
+    assert_no_selector :alert, text: "Message Created."
+
+    within_fieldset "Validate" do
+      select "published", from: "Status"
+      fill_in "Subject", with: "valid"
+      fill_in "Content", with: "valid"
+
+      assert_no_field valid: false
+
+      click_on "Create Message"
+    end
+
+    assert_selector :alert, text: "Message Created."
   end
 end